[REFACTOR] InputView 요청에 따라 개발, 입력한 결과 값 전달할 VO 객체 추가
