+++ src/main/java/view/vo/RaceConditionInput.java
@@ -1,4 +1,4 @@
-package view;
+package view.vo;
 
 public class RaceConditionInput {
